requirements.txt
boto3==1.37.1
mysqlclient==2.2.4
pandas==2.2.3
PyYAML==6.0.2